+++ packages/granularity/src/components/GrSelect/index.ts
@@ -11,6 +11,7 @@ export type {
   GrSelectOptionsView,
   GrSelectProps,
   GrSelectSize,
+  GrSelectState,
   GrSelectUnderline,
   GrSelectVariant,
   GrSelectView,